Blackwell 4: Blackwell Deception — Syntax update
This update fixes icons handling.

+++ play.it-2/games/play-blackwell-4-blackwell-deception.sh
@@ -1,8 +1,9 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
 # Copyright (c) 2015-2020, Antoine "vv221/vv222" Le Gonidec
+# Copyright (c) 2016-2020, Mopi
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,43 +31,42 @@ set -o errexit
 
 ###
 # Blackwell 4: Blackwell Deception
-# build native Linux packages from the original installers
-# send your bug reports to vv221@dotslashplay.it
+# build native packages from the original installers
+# send your bug reports to contact@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20200830.1
 
 # Set game-specific variables
 
 GAME_ID='blackwell-4-blackwell-deception'
 GAME_NAME='Blackwell 4: Blackwell Deception'
 
-ARCHIVES_LIST='ARCHIVE_GOG'
+ARCHIVES_LIST='
+ARCHIVE_GOG_0'
 
-ARCHIVE_GOG='gog_blackwell_deception_2.0.0.3.sh'
-ARCHIVE_GOG_URL='https://www.gog.com/game/blackwell_bundle'
-ARCHIVE_GOG_MD5='1c2684697a98eee4d64c7e34311fac6c'
-ARCHIVE_GOG_SIZE='610000'
-ARCHIVE_GOG_VERSION='1.0-gog2.0.0.3'
+ARCHIVE_GOG_0='gog_blackwell_deception_2.0.0.3.sh'
+ARCHIVE_GOG_0_URL='https://www.gog.com/game/blackwell_bundle'
+ARCHIVE_GOG_0_MD5='1c2684697a98eee4d64c7e34311fac6c'
+ARCHIVE_GOG_0_SIZE='610000'
+ARCHIVE_GOG_0_VERSION='1.0-gog2.0.0.3'
 
 ARCHIVE_DOC_DATA_PATH='data/noarch/docs'
-ARCHIVE_DOC_DATA_FILES='./*'
+ARCHIVE_DOC_DATA_FILES='*'
 
 ARCHIVE_GAME_BIN32_PATH='data/noarch/game'
-ARCHIVE_GAME_BIN32_FILES='./*.x86 ./lib'
+ARCHIVE_GAME_BIN32_FILES='*.x86 lib'
 
 ARCHIVE_GAME_BIN64_PATH='data/noarch/game'
-ARCHIVE_GAME_BIN64_FILES='./*.x86_64 ./lib64'
+ARCHIVE_GAME_BIN64_FILES='*.x86_64 lib64'
 
 ARCHIVE_GAME_DATA_PATH='data/noarch/game'
-ARCHIVE_GAME_DATA_FILES='./agsgame.dat ./acsetup.cfg ./audio.vox ./prog.bwd'
+ARCHIVE_GAME_DATA_FILES='agsgame.dat acsetup.cfg audio.vox prog.bwd'
 
 APP_MAIN_TYPE='native'
 APP_MAIN_EXE_BIN32='Deception.bin.x86'
 APP_MAIN_EXE_BIN64='Deception.bin.x86_64'
-APP_MAIN_ICONS_LIST='APP_MAIN_ICON'
 APP_MAIN_ICON='data/noarch/support/icon.png'
-APP_MAIN_ICON_RES='256'
 
 
 PACKAGES_LIST='PKG_BIN32 PKG_BIN64 PKG_DATA'
@@ -82,48 +82,55 @@ PKG_BIN64_DEPS="$PKG_BIN32_DEPS"
 
 # Load common functions
 
-target_version='2.4'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
+prepare_package_layout
 
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC_${PKG#PKG_}" "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}" "$PATH_GAME"
-done
+# Get game icon
 
 PKG='PKG_DATA'
-get_icon_from_temp_dir 'APP_MAIN'
+icons_get_from_workdir 'APP_MAIN'
+
+# Clean up temporary files
 
 rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
 # Write launchers
 
 for PKG in 'PKG_BIN32' 'PKG_BIN64'; do
-	write_launcher 'APP_MAIN'
+	launchers_write 'APP_MAIN'
 done
 
 # Build package
 
-postinst_icons_linking 'APP_MAIN'
-write_metadata 'PKG_DATA'
-write_metadata 'PKG_BIN32' 'PKG_BIN64'
+write_metadata
 build_pkg
 
 # Clean up
@@ -132,10 +139,6 @@ rm --recursive "$PLAYIT_WORKDIR"
 
 # Print instructions
 
-printf '\n'
-printf '32-bit:'
-print_instructions 'PKG_DATA' 'PKG_BIN32'
-printf '64-bit:'
-print_instructions 'PKG_DATA' 'PKG_BIN64'
+print_instructions
 
 exit 0
